requirements.txt
Pillow 
django-decouple 
transliterate
pandas 
bs4 
selenium 
xlsxwriter
django-filebrowser-no-grappelli 

django
django-cors-headers
django-autocomplete-light 
django-admin-autocomplete-filter
django-modeltranslation
django-mptt
django-tinymce4-lite 
# django-tinymce
django-ckeditor
django-filter
django-debug-toolbar
django-allauth
django-import-export
django-rosetta
django-colorfield
django-admin-sortable2
django-rest-framework
django-admin-rangefilter
django-debug-toolbar
django-nested-admin

docutils
pytz
feedparser
gunicorn
django-smtp-ssl
google-api-python-client
oauth2client






